Split partial laws into own trait.
