+++ scalacheck-binding/src/main/scala/spire/laws/GroupLaws.scala
@@ -21,46 +21,6 @@ trait GroupLaws[A] extends Laws {
   implicit def Equ: Eq[A]
   implicit def Arb: Arbitrary[A]
 
-
-  // partial groups
-
-  def semigroupoid(implicit A: Semigroupoid[A]) = new GroupProperties(
-    name = "semigroupoid",
-    parent = None,
-    "associative: a |+|?? b && b |+|?? c imply (a |+| b) |+|?? c" → forAll((a: A, b: A, c: A) =>
-      !((a |+|?? b) && (b |+|?? c)) || ((a |+|? b).get |+|?? c)
-    ),
-
-    "associative: (a |+|? b) |+|? c === a |+|? (b |+|? c)" → forAll((a: A, b: A, c: A) => {
-      (!(a |+|?? b) || !(b |+|?? c)) ||
-      ((a |+|? b).get |+|? c).get === (a |+|? (b |+|? c).get).get
-    }
-    )
-  )
-
-  def groupoid(implicit A: Groupoid[A]) = new GroupProperties(
-    name = "groupoid",
-    parent = Some(semigroupoid),
-    "left identity" → forAll((a: A) =>
-      (a.leftId |+|?? a) && ((a.leftId() |+|? a).get === a)
-    ),
-
-    "right identity" → forAll((a: A) =>
-      (a |+|?? a.rightId) && ((a |+|? a.rightId).get === a)
-    ),
-
-    "product with inverse is always defined" → forAll((a: A) =>
-      (a |+|?? a.inverse) && (a.inverse |+|?? a)
-    ),
-
-    "product with inverse is a left and right identity" → forAll((a: A, b: A) =>
-      !(a |+|?? b) || (
-        ((a |+|? b).get |+|? b.inverse).get === a &&
-          ((a.inverse |+|? a).get |+|? b).get === b
-      )
-    )
-  )
-
   // groups
 
   def semigroup(implicit A: Semigroup[A]) = new GroupProperties(
